+++ native/common/include/jp_longtype.h
@@ -64,12 +64,12 @@ public:
 
 	virtual jlong getAsLong(jvalue v) override
 	{
-		return field(v);
+		return (jlong) field(v);
 	}
 
 	virtual jdouble getAsDouble(jvalue v) override
 	{
-		return field(v);
+		return (jdouble) field(v);
 	}
 
 private:
